another failed attempt: 2023 day 1 part 2
I have tried to parse the broken word numbers from left to right
priority one by one, however it seems the priority should depend
on whether the word number is closer to the right or to the left,
which makes this approach even more complicated. There is a better way.